Updated State of Bugs/Unfinished Features
Offered candid commentary on what bugs/features were left unfinished. 
Documentation was a weak point for us in the final stretch of getting this thing working and submitted.
If anyone remembers what did/didn't work, or feels inclined to boot up the program themselves and update the README, feel free.